Moved global actions to ApplicationActions
