+++ app/features/application/ApplicationScreen.js
@@ -1,9 +1,8 @@
-import { Keyboard } from 'react-native';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 
 import * as ApiActionCreators from 'app/redux/shared_actions/ApiActionCreators';
-import * as SetActionCreators from 'app/redux/shared_actions/SetActionCreators';
+import * as ApplicationActions from './ApplicationActions';
 import ApplicationView from './ApplicationView';
 
 const mapStateToProps = (state) => ({
@@ -12,11 +11,7 @@ const mapStateToProps = (state) => ({
 
 const mapDispatchToProps = (dispatch) => {
     return bindActionCreators({
-        onChangeTab: () => (dispatch) => {
-            dispatch(ApiActionCreators.syncData());
-            dispatch(SetActionCreators.endOldWorkout());
-            Keyboard.dismiss();
-        },
+        onChangeTab: ApplicationActions.onChangeTab,
         onMount: ApiActionCreators.syncData
     }, dispatch);
 };
